+++ CHANGELOG.rst
@@ -76,6 +76,9 @@ Bug Fixes
 - Ensure varargin is correctly passed through to ``ndbase.pso`` from
   ``optmagk``. Previously user provided ``TolFun``, ``TolX`` and
   ``MaxIter`` would be overwritten by the defaults.
+- Warn users that that the results of ``spinwave`` have not been
+  scientifically validated for supercell structures with an
+  incommensurate modulation.
 - Emit warning if wrong length ``xmin``, ``xmax`` or ``x0`` is passed to
   ``optmagstr``. Previously they would be silently ignored.
 - No longer require a magnetic structure be initialised with ``genmagstr``